requirements.txt
requests
pandas
talib-binary
scikit-learn
backtrader
sqlalchemy
boto3
flask
telegram
pytest